Implemented new retrieve, plus multi-store & mult-retrieve.  test_retrieve_ts.sql and test_retrieve_ts.py exercise the new functionality, directly to the database and through JDBC, respecively
[git-p4: depot-paths = "//wcdba/dev/oracle/dev/": change = 19282]

+++ test/scripts/test_retrieve_ts.sql
@@ -1,8 +1,8 @@
 set serveroutput on;
 
 declare
-   l_tsinfo     timeseries_array := timeseries_array();
-   l_tsdata     tsv_array        := tsv_array();
+   l_tsinfo     cwms_20.timeseries_array := cwms_20.timeseries_array();
+   l_tsdata     cwms_20.tsv_array        := cwms_20.tsv_array();
    
    l_tsid       varchar2(183)  := 'Test-storeTs.Flow.Inst.1Hour.0.test';
    l_units      varchar2(16)   := 'cfs';
@@ -13,12 +13,15 @@ declare
    l_date_time  date;
    l_value      binary_double;
    l_quality    number;
-   l_tsrequest  timeseries_req_array := timeseries_req_array();
+   l_tsrequest  cwms_20.timeseries_req_array := cwms_20.timeseries_req_array();
    l_sequence   integer;
    i            pls_integer;
    l_ts_begin   timestamp;
    l_ts_end     timestamp;
    l_intvl      interval day to second;
+   l_time_zone  varchar2(28);
+   l_delete     boolean;
+   l_store      boolean;
 begin
    dbms_output.put_line('==>Creating data at ' || systimestamp);
    --
@@ -32,7 +35,7 @@ begin
          null;
       else
          l_tsdata.extend;
-         l_tsdata(l_tsdata.last) := tsv_type(from_tz(l_date_time, 'UTC'), i, 0);
+         l_tsdata(l_tsdata.last) := cwms_20.tsv_type(from_tz(l_date_time, 'UTC'), i, 0);
          if i > 20 then
             l_tsdata(l_tsdata.last).quality_code := 5;
          end if; 
@@ -41,38 +44,50 @@ begin
       l_date_time := l_date_time + 1 / 24;
    end loop;
    l_tsinfo.extend(4);
-   l_tsinfo(1) := timeseries_type('Tester-id1.Flow.Inst.0.0.test',     'kcfs', l_tsdata);
-   l_tsinfo(2) := timeseries_type('Tester-id2.Flow.Inst.1Hour.0.test', 'kcfs', l_tsdata);
-   l_tsinfo(3) := timeseries_type('Tester-id3.Flow.Inst.0.0.test',     'cms',  l_tsdata);
-   l_tsinfo(4) := timeseries_type('Tester-id4.Flow.Inst.1Hour.0.test', 'cms',  l_tsdata);
+   l_tsinfo(1) := cwms_20.timeseries_type('Tester-id1.Flow.Inst.0.0.test',     'kcfs', l_tsdata);
+   l_tsinfo(2) := cwms_20.timeseries_type('Tester-id2.Flow.Inst.1Hour.0.test', 'kcfs', l_tsdata);
+   l_tsinfo(3) := cwms_20.timeseries_type('Tester-id3.Flow.Inst.0.0.test',     'cms',  l_tsdata);
+   l_tsinfo(4) := cwms_20.timeseries_type('Tester-id4.Flow.Inst.1Hour.0.test', 'cms',  l_tsdata);
+   l_tsrequest.extend(l_tsinfo.count);
+   
+   for i in 1..l_tsinfo.count loop
+      l_tsrequest(i) := cwms_20.timeseries_req_type(l_tsinfo(i).tsid, l_units, l_start_time, l_end_time);
+   end loop;
+   l_tsrequest(4).unit := null;
    --
    -- delete any existing data
    --
-   l_ts_begin := systimestamp;
-   dbms_output.put_line('==>Starting delete at ' || l_ts_begin);
-   for i in 1..l_tsinfo.count loop
-      begin
-         cwms_ts.delete_ts(l_tsinfo(i).tsid, cwms_util.delete_ts_cascade, 'NAB');
-      exception
-         when others then
-            if instr(sqlerrm, 'TS_ID_NOT_FOUND') != 0 then
-               null;
-            end if;
-      end;
-   end loop;
-   l_ts_end := systimestamp;
-   l_intvl  := l_ts_end - l_ts_begin; 
-   dbms_output.put_line('==>Delete done at ' || l_ts_end || ' (' || l_intvl || ')');
+   l_delete := false;
+   if l_delete then
+      l_ts_begin := systimestamp;
+      dbms_output.put_line('==>Starting delete at ' || l_ts_begin);
+      for i in 1..l_tsinfo.count loop
+         begin
+            cwms_ts.delete_ts(l_tsinfo(i).tsid, cwms_util.delete_ts_cascade, 'NAB');
+         exception
+            when others then
+               if instr(sqlerrm, 'TS_ID_NOT_FOUND') != 0 then
+                  null;
+               end if;
+         end;
+      end loop;
+      l_ts_end := systimestamp;
+      l_intvl  := l_ts_end - l_ts_begin; 
+      dbms_output.put_line('==>Delete done at ' || l_ts_end || ' (' || l_intvl || ')');
+   end if;
    --
    -- store the data
    --
-   l_ts_begin := systimestamp;
-   dbms_output.put_line('==>Starting multi-store at ' || l_ts_begin);
-   cwms_ts.store_ts_multi(l_tsinfo,cwms_util.delete_insert,'F',cwms_util.non_versioned,'NAB');
-   commit;
-   l_ts_end := systimestamp;
-   l_intvl  := l_ts_end - l_ts_begin; 
-   dbms_output.put_line('==>Multi-store done at ' || l_ts_end || ' (' || l_intvl || ')');
+   l_store := true;
+   if l_store then
+      l_ts_begin := systimestamp;
+      dbms_output.put_line('==>Starting multi-store at ' || l_ts_begin);
+      cwms_ts.store_ts_multi(l_tsinfo,cwms_util.delete_insert,'F',cwms_util.non_versioned,'NAB');
+      commit;
+      l_ts_end := systimestamp;
+      l_intvl  := l_ts_end - l_ts_begin; 
+      dbms_output.put_line('==>Multi-store done at ' || l_ts_end || ' (' || l_intvl || ')');
+   end if;
    --
    -- retrieve the data in a loop
    --
@@ -86,15 +101,15 @@ begin
       cwms_ts.retrieve_ts(
          l_cursor,
          l_tsinfo(i).tsid,
-         l_units,
+         l_tsrequest(i).unit,
          l_start_time,
          l_end_time,
-         'US/Central', -- time zone
+         'UTC', -- time zone
          'F',          -- trim
          'T',          -- start inclusive
          'T',          -- end inclusive
-         'T',          -- previous
-         'T',          -- next
+         'F',          -- previous
+         'F',          -- next
           null,        -- version date
          'T',          -- max version
          'NAB');       -- office id
@@ -123,24 +138,20 @@ begin
    l_intvl  := l_ts_end - l_ts_begin; 
    dbms_output.put_line('==>Retrieve loop done at ' || l_ts_end || ' (' || l_intvl || ')');
    --
-   -- retrieve the data with retrive_ts2_multi
+   -- retrieve the data with retrive_ts_multi
    --
    dbms_output.put_line('-----------------------------------------');
-   l_tsrequest.extend(l_tsinfo.count);
-   for i in 1..l_tsinfo.count loop
-      l_tsrequest(i) := timeseries_req_type(l_tsinfo(i).tsid, l_units, l_start_time, l_end_time);
-   end loop;
    l_ts_begin := systimestamp;
    dbms_output.put_line('==>Starting multi-retrieve at ' || l_ts_begin);
    cwms_ts.retrieve_ts_multi(
       l_cursor,
       l_tsrequest,
-      'US/Central', -- time zone
+      'UTC', -- time zone
       'F',     -- trim
       'T',     -- start inclusive
       'T',     -- end inclusive
-      'T',     -- previous
-      'T',     -- next
+      'F',     -- previous
+      'F',     -- next
       null,    -- version date
       'T',     -- max version
       'NAB');  -- office
@@ -148,26 +159,27 @@ begin
    l_intvl  := l_ts_end - l_ts_begin; 
    dbms_output.put_line('==>Multi-retrieve done at ' || l_ts_end || ' (' || l_intvl || ')');
    loop
-      fetch l_cursor into l_sequence, l_tsid, l_units, l_start_time, l_end_time, l_cursor2;
+      fetch l_cursor into l_sequence, l_tsid, l_units, l_start_time, l_end_time, l_time_zone, l_cursor2;
       exit when l_cursor%notfound;
       dbms_output.put_line( 
-         ''  || l_sequence 
+         ''  || l_sequence
              || ' ' || l_tsid 
              || ' ' || l_units
              || ' ' || to_char(l_start_time, 'yyyy/mm/dd-hh24:mi:ss')
-             || ' ' || to_char(l_end_time, 'yyyy/mm/dd-hh24:mi:ss'));
+             || ' ' || to_char(l_end_time, 'yyyy/mm/dd-hh24:mi:ss')
+             || ' ' || l_time_zone);
       loop
          fetch l_cursor2 into l_date_time, l_value, l_quality;
          exit when l_cursor2%notfound;
          if l_value is null then
             dbms_output.put_line(
-               '----> ' 
+               chr(9) 
                || to_char(l_date_time, 'yyyy/mm/dd-hh24:mi:ss') 
                || ' <null> ' 
                || l_quality);
          else
             dbms_output.put_line(
-               '----> ' 
+               chr(9) 
                || to_char(l_date_time, 'yyyy/mm/dd-hh24:mi:ss') 
                || ' ' 
                || l_value
